Test that uses assert_cmd requires to be outside the main module to match with non harcoded folder structures, move them to tests folder looks like solve the issue, thanks @epage for the advice fix #1
