+++ src/args.rs
@@ -4,7 +4,9 @@ use clap::Parser;
 #[clap(name = "say")]
 #[clap(version = "v0.1.0")]
 #[clap(author = "CarlosEduardoL")]
-#[clap(about = "Rust echo copy", long_about = "If -e is in effect, the following sequences are recognized:
+#[clap(
+    about = "Rust echo copy",
+    long_about = "If -e is in effect, the following sequences are recognized:
   \\\\     backslash
   \\a     alert (BEL)
   \\b     backspace
@@ -16,7 +18,8 @@ use clap::Parser;
   \\t     horizontal tab
   \\v     vertical tab
   \\0NNN  byte with octal value NNN (1 to 3 digits)
-  \\xHH   byte with hexadecimal value HH (1 to 2 digits)")]
+  \\xHH   byte with hexadecimal value HH (1 to 2 digits)"
+)]
 pub struct SayArgs {
     #[clap(value_name = "TEXT")]
     /// Text to be printed on the stdout
@@ -26,5 +29,5 @@ pub struct SayArgs {
     pub trim_new_line: bool,
     #[clap(short = 'e')]
     /// enable interpretation of backslash escapes
-    pub backslash_interpretation: bool
+    pub backslash_interpretation: bool,
 }
